+++ tests/test_Encoder.py
@@ -9,6 +9,6 @@ def test_encoding_caesar():
 
 def test_encoding_custom():
     coder = Coder()
-    result = coder.encode("password 1","hello world", EncryptionType.CUSTOM_CIPHER, custom_salt="abcdef") # TODO: need a way to force constant salt
+    result = coder.encode("password 1","hello world", EncryptionType.CUSTOM_CIPHER, custom_salt="abcdefgh") # TODO: need a way to force constant salt
 
-    assert result == 'qiAgG10TXfvzJd5oliiRWWyZ3v8hnYgdSGzgGrhzmSexje1KKc8qiB76WENHCTjEct/XbiHWZfUvPGiui12n4FN7l0zsh9ijabcdef'
+    assert result == "3LIT0VjSogGTnydfZlz9hmuAoDiPDNFucny0E6J8Pnput1b+rTW1u/GcC9TdOaU/PBOhpFJ6RRItwCEE5MVe01N7l0zsh9ijabcdefgh"